Improve delete confirmation dialogs

+++ mico-admin/src/app/app-detail/app-detail.component.ts
@@ -27,6 +27,7 @@ import { versionComparator } from '../api/semantic-version';
 import { CreateNextVersionComponent } from '../dialogs/create-next-version/create-next-version.component';
 import { MatDialog, MatSnackBar } from '@angular/material';
 import { safeUnsubscribe } from '../util/utils';
+import { YesNoDialogComponent } from '../dialogs/yes-no-dialog/yes-no-dialog.component';
 
 @Component({
     selector: 'mico-app-detail',
@@ -261,15 +262,29 @@ export class AppDetailComponent implements OnInit, OnDestroy {
      * removes the current application version
      */
     deleteCurrentVersion() {
-        this.apiService.deleteApplication(this.application.shortName, this.selectedVersion)
-            .subscribe(val => {
 
-                // stay on the application page if there exists another version
-                if (this.allVersions.length > 0) {
-                    this.updateVersion(null);
-                } else {
-                    this.router.navigate(['../app-list']);
-                }
-            });
+        const dialogRef = this.dialog.open(YesNoDialogComponent, {
+            data: {
+                object: { name: this.application.name, shortName: this.shortName, version: this.selectedVersion },
+                question: 'deleteApplication'
+            }
+        });
+
+        const subDeleteDependency = dialogRef.afterClosed().subscribe(result => {
+            if (result) {
+
+                this.apiService.deleteApplication(this.application.shortName, this.selectedVersion)
+                    .subscribe(val => {
+
+                        // stay on the application page if there exists another version
+                        if (this.allVersions.length > 0) {
+                            this.updateVersion(null);
+                        } else {
+                            this.router.navigate(['../app-list']);
+                        }
+                    });
+                safeUnsubscribe(subDeleteDependency);
+            }
+        });
     }
 }
